pyproject.toml
[project]
name = "enemane-ai"
version = "0.1.0"
description = "Streamlit app to batch-analyze graph images and PDFs with preset prompts."
readme = "README.md"
requires-python = "==3.13.*"
license = { text = "MIT" }
authors = [{ name = "Your Name" }]
dependencies = [
    "streamlit>=1.33,<2",
    "pillow>=10.3",
    "pypdf>=4.0",
    "pypdfium2>=4.30",
    "google-genai>=1.0,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-mock>=3.14",
    "pre-commit>=3.7",
    "ruff>=0.6.9",
    "mypy>=1.12",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.2",
    "pytest-mock>=3.14",
    "pre-commit>=3.7",
    "ruff>=0.6.9",
    "mypy>=1.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "W",
    "B",
    "C4",
    "FAST",
    "PERF",
    "RUF",
]
ignore = [
    "E203",
    "E266",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.13"
check_untyped_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true
